Refactor packages for better structure
