--- src/main/java/com/example/geektrust/PathFindingStrategy.java
@@ -1,5 +1,0 @@
-package com.example.geektrust;
-
-public interface PathFindingStrategy {
-    int calculatePower(Position source, Position destination, Direction direction);
-}
