+++ src/main/java/com/example/geektrust/command/CommandContext.java
@@ -1,4 +1,8 @@
-package com.example.geektrust;
+package com.example.geektrust.command;
+
+import com.example.geektrust.model.Board;
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.service.PathFindingStrategy;
 
 public class CommandContext {
     private int sourceX = 0;
