+++ src/main/java/com/example/geektrust/command/CommandFactory.java
@@ -1,4 +1,7 @@
-package com.example.geektrust;
+package com.example.geektrust.command;
+
+import com.example.geektrust.model.Board;
+import com.example.geektrust.model.Direction;
 
 class CommandFactory {
     private static final int SOURCE_ARGS_COUNT = 4;
