+++ src/main/java/com/example/geektrust/model/GMan.java
@@ -1,4 +1,4 @@
-package com.example.geektrust;
+package com.example.geektrust.model;
 
 public class GMan {
     private Position position;
