+++ src/main/java/com/example/geektrust/model/State.java
@@ -1,4 +1,4 @@
-package com.example.geektrust;
+package com.example.geektrust.model;
 
 public class State {
     private final Position position;
