+++ src/main/java/com/example/geektrust/service/PathFindingStrategy.java
@@ -1,0 +1,8 @@
+package com.example.geektrust.service;
+
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.model.Position;
+
+public interface PathFindingStrategy {
+    int calculatePower(Position source, Position destination, Direction direction);
+}
