+++ src/test/java/com/example/geektrust/GManTest.java
@@ -1,5 +1,8 @@
 package com.example.geektrust;
 
+import com.example.geektrust.model.Board;
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.model.GMan;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
