+++ src/test/java/com/example/geektrust/MainTest.java
@@ -1,5 +1,8 @@
 package com.example.geektrust;
 
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.model.Position;
+import com.example.geektrust.service.PowerCalculator;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
