+++ src/test/java/com/example/geektrust/PositionTest.java
@@ -1,5 +1,7 @@
 package com.example.geektrust;
 
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.model.Position;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
