+++ src/test/java/com/example/geektrust/SourceCommandTest.java
@@ -1,5 +1,10 @@
 package com.example.geektrust;
 
+import com.example.geektrust.command.CommandContext;
+import com.example.geektrust.command.SourceCommand;
+import com.example.geektrust.model.Direction;
+import com.example.geektrust.model.Position;
+import com.example.geektrust.service.PathFindingStrategy;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
